chore: restructure files in pipeline-viewport.
