+++ services/orchest-webserver/client/src/pipeline-view/pipeline-viewport/NoStep.tsx
@@ -1,6 +1,6 @@
 import { StackProps } from "@mui/material/Stack";
 import React from "react";
-import { ViewportCenterMessage } from "./ViewportCenterMessage";
+import { ViewportCenterMessage } from "./components/ViewportCenterMessage";
 
 export const NoStep = (props: StackProps) => {
   return (
