+++ services/orchest-webserver/client/src/pipeline-view/pipeline-viewport/components/Overlay.tsx
@@ -1,0 +1,6 @@
+import Box from "@mui/material/Box";
+import React from "react";
+
+export const Overlay = () => (
+  <Box sx={{ width: "100vw", height: "100vh", display: "block" }} />
+);
